# GoChatbotBackend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is an API for managing room bookings.",
        "title": "Booking API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
        "/api/bookings": {
            "post": {
                "description": "Add a new booking for a specific room and timeslot",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "text/plain"
                ],
                "tags": [
                    "bookings"
                ],
                "summary": "Create a booking",
                "parameters": [
                    {
                        "description": "Booking info",
                        "name": "booking",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.Booking"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Booking created",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Failed to insert booking",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/rooms": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "rooms"
                ],
                "summary": "Get all rooms",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "object",
                                "additionalProperties": true
                            }
                        }
                    },
                    "500": {
                        "description": "Failed to fetch rooms",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "handlers.Booking": {
            "type": "object",
            "properties": {
                "date": {
                    "description": "format: \"YYYY-MM-DD\"",
                    "type": "string"
                },
                "room_id": {
                    "type": "integer"
                },
                "timeslot_id": {
                    "type": "integer"
                }
            }
        }
    }
}